Introduce named global executor for choice in execution (#2)
It is a good for binary crate to active at most one global spawner
compat. But it could be relatively hard if dependency graph is large.
In this case, it would be good to offer choices in execution.

This commit uses environment variable SPAWNS_GLOBAL_EXECUTOR to choose
one in absent of thread context spawners.